search/requirements.txt
mcp>=1.2.0
fastmcp>=0.4.1
azure-search-documents>=11.4.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
uvicorn>=0.22.0